+++ _src/managers/consumers/nodejsHttp.ts
@@ -176,7 +176,7 @@ export class NaniumConsumerNodejsHttp implements ServiceManager {
 									deserialized = this.config.serializer.deserializePartial(chunk.toString(), restFromLastTime);
 									if (deserialized.data?.length) {
 										for (const data of deserialized.data) {
-											observer.next(NaniumObject.plainToClass(
+											observer.next(NaniumObject.create(
 												data,
 												request.constructor[responseTypeSymbol],
 												request.constructor[genericTypesSymbol]