+++ _src/tests/testLogger.ts
@@ -3,9 +3,9 @@ import { Logger, LogLevel } from '../interfaces/logger';
 export class TestLogger implements Logger {
 	loglevel: LogLevel;
 
-	errors: any[] = [];
-	warnings: any[] = [];
-	infos: any[] = [];
+	errors: any[][] = [];
+	warnings: any[][] = [];
+	infos: any[][] = [];
 
 	constructor(level: LogLevel) {
 		this.loglevel = level;
@@ -13,19 +13,19 @@ export class TestLogger implements Logger {
 
 	error(...args: any[]): void {
 		if (this.loglevel >= LogLevel.error) {
-			args.forEach(arg => this.errors.push(arg));
+			this.errors.push(args);
 		}
 	}
 
 	warn(...args: any[]): void {
 		if (this.loglevel >= LogLevel.warn) {
-			args.forEach(arg => this.warnings.push(arg));
+			this.warnings.push(args);
 		}
 	}
 
 	info(...args: any[]): void {
 		if (this.loglevel >= LogLevel.info) {
-			args.forEach(arg => this.infos.push(arg));
+			this.infos.push(args);
 		}
 	}
 }